refactor: update avatar images
